Refactor: 웹툰 컴포넌트 관련 디렉토리 이름, 파일이름, 변수이름 변경
