+++ js/main/dayFilter/replaceDayContents.js
@@ -1,4 +1,4 @@
-import webtoonComponent from '../webtoonComponent/webtoonComponent.js';
+import createWebtoonComponent from '../webtoon/createComponent.js';
 
 const replaceDayContents = ({ webtoonArr, coverType }) => {
   const contentsContainer = document.querySelector('.contents-container');
@@ -12,8 +12,8 @@ const replaceDayContents = ({ webtoonArr, coverType }) => {
   webtoonComponentContainer.className = 'webtoon-component-container';
 
   for (let i = 0; i < webtoonArr.length; i++) {
-    const webtoonEl = webtoonComponent(coverType, webtoonArr[i]);
-    webtoonComponentContainer.appendChild(webtoonEl);
+    const $webtoonComponent = createWebtoonComponent(coverType, webtoonArr[i]);
+    webtoonComponentContainer.appendChild($webtoonComponent);
   }
   dayContentsMain.appendChild(webtoonComponentContainer);
   contentsContainer.appendChild(dayContentsMain);
