+++ js/main/webtoon/coverTemplete.js
